+++ src/pages/two-player-home/two-player-home.component.tsx
@@ -3,10 +3,12 @@ import './two-player-home.styles.scss';
 import React, { FC } from 'react';
 
 import GlobalButton from '../../components/global-button/global-button.component';
+import MainGameScreen from '../../components/main-game-screen/main-game-screen.component';
 
 const TwoPlayerHome: FC = () => {
   return (
     <div className='global-page-container'>
+      <div className='pixelated-grid'></div>
       <div className='two-player-home-container'>
         <div className='two-player-home-back-button'>
           <GlobalButton label='<- Back' route='/' />
@@ -15,8 +17,7 @@ const TwoPlayerHome: FC = () => {
           <h1 className='header-text'>2 Player</h1>
         </header>
         <div className='two-player-home-content'>
-          {/* <TicTacToe /> */}
-          <p>In Progress...</p>
+          <MainGameScreen />
         </div>
       </div>
     </div>