+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 - Added Main Menu and initial page setup
   - Added initial Single Player page
   - Added initial 2 Player page
+- Updated global background (pixelated grid) and button styles
+- Added main game screen page layout and style (no functionality yet)